create records

+++ prelab2/prelab2/UserManagement.cs
@@ -56,47 +56,61 @@ namespace ooplab
         }
         private void btnsave_Click(object sender, EventArgs e)
         {
-            var csv = new StringBuilder();
-            if (txtnewpassword.Text == txtconfirm.Text)
+            if (Form1.Loaduser.Type == "Admin")
             {
+                var csv = new StringBuilder();
+  
                 File.Delete(@"Data\user.csv");
                 string fileName = @"Data\user.csv";
                 File.Create(fileName).Close();
-                users.Userlist[i].Password = Hash256.ComputeSha256Hash(txtnewpassword.Text);
-                for (int j = 0; j < users.Userlist.Count; j++)
+                if (txtnewpassword.Text == txtconfirm.Text)
                 {
+                    users.Userlist[i].Password = Hash256.ComputeSha256Hash(txtnewpassword.Text);
+                    for (int j = 0; j < users.Userlist.Count; j++)
                     {
-                        var newLine = string.Format("{0};{1};{2}", users.Userlist[j].Username, users.Userlist[j].Password, users.Userlist[j].Type, Environment.NewLine);
-                        csv.AppendLine(newLine);
-                        lblmassage.Text = "Success";
-                        txtList.Text = "Username: " + users.Userlist[i].Username + Environment.NewLine + "Password: " +
-                           users.Userlist[i].Password + Environment.NewLine + "Type: " + users.Userlist[i].Type;
+                        {
+                            var newLine = string.Format("{0};{1};{2};{3}", users.Userlist[j].Username, users.Userlist[j].Password,
+                                users.Userlist[j].Type,users.Userlist[j].Phone_number, Environment.NewLine);
+                            csv.AppendLine(newLine);
+                            lblmassage.Text = "Success";
+                            txtList.Text = "Username: " + users.Userlist[i].Username + Environment.NewLine + "Password: " +
+                               users.Userlist[i].Password + Environment.NewLine + "Type: " + users.Userlist[i].Type;
+                        }
                     }
+                    File.AppendAllText(@"Data\user.csv", csv.ToString());
+                    lblmassage.Text = "Accsess";
                 }
-                File.AppendAllText(@"Data\user.csv", csv.ToString());
+                else
+                    lblmassage.Text = "Not Confirm";
             }
             else
-                lblmassage.Text = "Not Confirm";
+                lblmassage.Text = "You can't do this, you are not an Admin";              
+            
         }
 
         private void btnSaveType_Click(object sender, EventArgs e)
         {
-            var csv = new StringBuilder();
-            File.Delete(@"Data\user.csv");
-            string fileName = @"Data\user.csv";
-            File.Create(fileName).Close();
-            users.Userlist[i].Type = cmbNewType.SelectedItem.ToString();
-            for (int j = 0; j < users.Userlist.Count; j++)
+            if (Form1.Loaduser.Type == "Admin")
             {
+                var csv = new StringBuilder();
+                File.Delete(@"Data\user.csv");
+                string fileName = @"Data\user.csv";
+                File.Create(fileName).Close();
+                users.Userlist[i].Type = cmbNewType.SelectedItem.ToString();
+                for (int j = 0; j < users.Userlist.Count; j++)
                 {
-                    var newLine = string.Format("{0};{1};{2}", users.Userlist[j].Username, users.Userlist[j].Password, users.Userlist[j].Type, Environment.NewLine);
-                    csv.AppendLine(newLine);
-                    lblmassage.Text = "Success";
-                    txtList.Text = "Username: " + users.Userlist[i].Username + Environment.NewLine + "Password: " +
-                           users.Userlist[i].Password + Environment.NewLine + "Type: " + users.Userlist[i].Type;
+                    {
+                        var newLine = string.Format("{0};{1};{2};{3}", users.Userlist[j].Username, users.Userlist[j].Password,
+                            users.Userlist[j].Type,users.Userlist[j].Phone_number, Environment.NewLine);
+                        csv.AppendLine(newLine);
+                        lblmassage.Text = "Success";
+                        txtList.Text = "Username: " + users.Userlist[i].Username + Environment.NewLine + "Password: " +
+                               users.Userlist[i].Password + Environment.NewLine + "Type: " + users.Userlist[i].Type;
+                    }
                 }
+                File.AppendAllText(@"Data\user.csv", csv.ToString());
+                lblmassage.Text = "Accsess";
             }
-            File.AppendAllText(@"Data\user.csv", csv.ToString());
         }
 
         private void lblminimize_Click(object sender, EventArgs e)
